Block partitioning support for D8, SparsePartition, and ARP

+++ src/d8.h
@@ -12,6 +12,6 @@ void writeSlope(tdpartition *flowDir, tdpartition *elevDEM, tdpartition* slopefi
 //Open files, initialize grid memory....
 int setdird8(char* demfile, char* pointfile, char *slopefile, char *flowfile, int useflowfile);
 
-long setPosDir(linearpart<float>& elevDEM, linearpart<short>& flowDir);
+int setPosDir(linearpart<float>& elevDEM, linearpart<short>& flowDir);
 
 #endif
